Added orderNumber field to ServiceOrder

+++ src/main/java/com/aprimore/models/dtos/PcpPdfDto.java
@@ -7,11 +7,11 @@ public class PcpPdfDto {
     
 
     private Integer sequencia;
-    private long numeroOS;
+    private int numeroOS;
     private String empresa;
     private String trabalho;
 
-    public PcpPdfDto(Integer sequencia, long numeroOS, String empresa, String trabalho) {
+    public PcpPdfDto(Integer sequencia, int numeroOS, String empresa, String trabalho) {
         this.sequencia = sequencia;
         this.numeroOS = numeroOS;
         this.empresa = empresa;
